pyproject.toml
[project]
name = "kyc-document-processing"
version = "0.1.0"
description = "Démonstrateur KYC - Traitement de documents avec LLM (RAD/LAD multi-classes)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform>=1.38.0",
    "google-cloud-documentai>=2.20.0",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.setuptools]
packages = ["src"]
